+++ dtool/cli.py
@@ -13,7 +13,17 @@ from dtool import (
     DataSet,
 )
 from dtool.clickutils import create_project, generate_descriptive_metadata
-from dtool.datatool import README_SCHEMA
+
+README_SCHEMA = [
+    ("project_name", u"project_name"),
+    ("dataset_name", u"dataset_name"),
+    ("confidential", False),
+    ("personally_identifiable_information", False),
+    ("owner_name", u"Your Name"),
+    ("owner_email", u"your.email@example.com"),
+    ("owner_username", u"namey"),
+    ("date", u"today"),
+]
 
 logger = sender.FluentSender('arctool', host='v0679', port=24224)